+++ test/packs/fake_v2.ts
@@ -123,6 +123,7 @@ pack.addDynamicSyncTable({
   description: 'FooDynDesc',
   getName: async () => 'name',
   getDisplayUrl: async () => 'display url',
+  identityName: 'FooDyn',
   getSchema: async () => {
     return {
       type: coda.ValueType.Object,